# temp-gigachat-provider/openapi/gigachat-openapi.yml
openapi: 3.0.3
info:
  title: GigaChat API
  version: 1.0.0
  contact:
    name: GigaChat API
    url: 'https://developers.sber.ru/portal/products/gigachat-api'
    email: gigachat@sberbank.ru
  description: |
  
    Справочная документация по REST API нейросетевой модели GigaChat.

    О стоимости и условиях использования GigaChat API вы можете узнать в разделе [Тарифы и оплата](/ru/gigachat/api/tariffs).

    ## Получение токена доступа и авторизация запросов

    Запросы к GigaChat API передаются по адресу `https://gigachat.devices.sberbank.ru/` и авторизуются с помощью токена доступа по протоколу [OAuth 2.0](https://tools.ietf.org/html/rfc6749).
    Токен доступа передается в заголовке `Authorization`:
    
    ```sh
    curl -L -X GET 'https://gigachat.devices.sberbank.ru/api/v1/models' \
    -H 'Accept: application/json' \
    -H 'Authorization: Bearer <токен_доступа>'
    ```

    :::tip
    
    Вы также можете передавать запросы к [моделям в раннем доступе](/ru/gigachat/models/preview-models).
    Их возможности могут отличаться от моделей, доступных в промышленном контуре.

    :::

    Чтобы получить токен, отправьте запрос # START_RAW<APIMethod type="POST" path="/api/v2/oauth" link="/ru/gigachat/api/reference/rest/post-token"/># END_RAW:

    ```sh
    curl -L -X POST 'https://ngw.devices.sberbank.ru:9443/api/v2/oauth' \
    -H 'Content-Type: application/x-www-form-urlencoded' \
    -H 'Accept: application/json' \
    -H 'RqUID: <идентификатор_запроса>' \
...
    ```
...
servers:
...
paths:
...
  /models:
    get:
      tags:
        - models
      x-codeSamples:
        - lang: "Python"
          label: "gigachat"
          source: |
            """
            Установите библиотеку gigachat с помощью менеджера пакетов pip:

            pip install gigachat
            """
            from gigachat import GigaChat

            giga = GigaChat(
               credentials="<ключ_авторизации>",
            )

            response = giga.get_models()

            print(response)
        - lang: "JavaScript"
          label: "gigachat"
          source: |
            /**
             * Установите библиотеку gigachat с помощью менеджера пакетов npm:
             * 
             * npm install gigachat
             */
            import GigaChat from 'gigachat';

            const giga = new GigaChat({
                credentials: '<ключ_авторизации>',
                scope: '<версия_API>',
            });

            const response = await giga.getModels();

            console.log(response);
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models'
          description: OK
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - Токен доступа: []
      operationId: getModels
      summary: Получить список моделей
      description: Возвращает массив объектов с данными доступных моделей. Описание доступных моделей в разделе [Модели GigaChat](/ru/gigachat/models).
  /files:
    get:
      tags:
        - files-storage
      x-codeSamples:
        - lang: "Python"
          label: "gigachat"
          source: |
            """
            Установите библиотеку gigachat с помощью менеджера пакетов pip:

            pip install gigachat
            """
            from gigachat import GigaChat

            giga = GigaChat(
               credentials="<ключ_авторизации>"
            )

            response = giga.get_files()

            print(response)
        - lang: "JavaScript"
          label: "gigachat"
          source: |
            /**
             * Установите библиотеку gigachat с помощью менеджера пакетов npm:
             * 
             * npm install gigachat
             */
            import GigaChat from 'gigachat';

            const giga = new GigaChat({
              credentials: '<ключ_авторизации>',
              scope: '<версия_API>',
            });

            const files = await giga.getFiles();

            console.log(files);
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Files'
          description: OK
      security:
        - Токен доступа: []
      operationId: getFiles
      summary: Получить список доступных файлов
      description: Возвращает массив объектов с данными доступных файлов.
    post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUpload'
        required: true
      tags:
        - files-storage
      x-codeSamples:
        - lang: "Python"
          label: "gigachat"
          source: |
            """
            Установите библиотеку gigachat с помощью менеджера пакетов pip:

            pip install gigachat
            """
            from gigachat import GigaChat

            giga = GigaChat(
               credentials="<ключ_авторизации>",
            )

            response = giga.upload_file(open("/<путь_к_файлу>/<имя_файла>.txt", mode="rb"))

            print(response)
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
          description: OK
      security:
        - Токен доступа: []
      operationId: postFile
      summary: Загрузить файл
      description: |
        Загружает в хранилище текстовые документы, изображения или аудиофайлы.
        Возвращает объект с данными загруженного файла.
        Загруженные файлы доступны только вам.

        Идентификатор файла, указанный в поле `id`, можно использовать при [запросах на генерацию](/ru/gigachat/api/reference/rest/post-chat).
        Для этого идентификаторы нужно передать в массиве `attachments`.
        Подробнее — в разделе [Обработка файлов](/ru/gigachat/guides/working-with-files).

        Хранилище поддерживает текстовые документы, изображения и аудиофайлы разных форматов.

        # START_RAW
        <Tabs queryString="ext">
        <TabItem value="text" label="Текстовые документы" default>
        # END_RAW

        ```mdx-code-block
        <APITable colsWidth={['200px']}>
        ```

        | Формат | MIME-тип   |
        |--------|------------|
        | txt    | text/plain  |
        | doc    | application/msword  |
        | docx   | application/vnd.openxmlformats-officedocument.wordprocessingml.document |
        | pdf    | application/pdf  |
        | epub   | application/epub  |
        | ppt    | application/ppt  |
        | pptx   | application/pptx  |

        ```mdx-code-block
        </APITable>
        ```
        
        # START_RAW
        </TabItem>
        <TabItem value="image" label="Изображения" >
        # END_RAW

        ```mdx-code-block
        <APITable colsWidth={['200px']}>
        ```
        
        | Формат | MIME-тип   |
        |--------|------------|
        | jpeg    | image/jpeg  |
        | png    | image/png  |
        | tiff   | image/tiff |
        | bmp    | image/bmp  |

        ```mdx-code-block
        </APITable>
        ```

        # START_RAW
        </TabItem>
        <TabItem value="audio" label="Аудиофайлы" >
        # END_RAW
        
        ```mdx-code-block
        <APITable colsWidth={['200px']}>
        ```
        
        | Формат | MIME-тип                                                       |
        | ------ | -------------------------------------------------------------- |
        | mp4    | audio/mp4                                                      |
        | mp3    | audio/mp3                                                      |
        | m4a    | audio/x-m4a                                                    |
        | wav    | audio/x-wav<br />audio/wave<br />audio/wav<br />audio/x-pn-wav |
        | weba   | audio/webm                                                |
        | ogg    | audio/x-ogg                                                    |
        | opus   | audio/opus                                                     |

        ```mdx-code-block
        </APITable>
        ```

        # START_RAW
        </TabItem>
        </Tabs>
        # END_RAW

        На размеры файлов действуют ограничения:

        * максимальный размер одного аудиофайла в запросе — 35 Мб;
        * максимальный размер одного изображения в запросе — 15 Мб;
        * максимальный размер одного текстового файла в запросе — 40 Мб.

        Общий размер загружаемых аудиофайлов и изображений должен быть меньше 80 Мб.

          :::note

          При использовании больших текстовых файлов в запросах на генерацию, их содержимое может превышать [размер контекста модели](/ru/gigachat/models#modeli-dlya-generatsii).
          В таком случае вернется [ошибка с кодом 422](/ru/gigachat/api/errors-description?responseCode=422).

          :::
...
  /chat/completions:
    post:
      parameters:
        - $ref: '#/components/parameters/xClientId'
        - $ref: '#/components/parameters/xRequestId'
        - $ref: '#/components/parameters/xSessionId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Chat'
              example: {"model": "GigaChat","messages": [{"role": "system","content": "Ты — профессиональный переводчик на английский язык. Переведи точно сообщение пользователя."},{"role": "user","content": "GigaChat — это сервис, который умеет взаимодействовать с пользователем в формате диалога, писать код, создавать тексты и картинки по запросу пользователя."}],"stream": false,"update_interval": 0}
      x-codeSamples:
        - lang: "Python"
          label: "gigachat"
          source: |
            """
            Установите библиотеку gigachat с помощью менеджера пакетов pip:

            pip install gigachat
            """
            from gigachat import GigaChat

            giga = GigaChat(
               credentials="<ключ_авторизации>",
            )

            response = giga.chat("Расскажи про себя")

            print(response)
        - lang: "JavaScript"
          label: "gigachat"
          source: |
            /**
             * Установите библиотеку gigachat с помощью менеджера пакетов npm:
             * 
             * npm install gigachat
             */
            import GigaChat from 'gigachat';

            const giga = new GigaChat({
              credentials: '<ключ_авторизации>',
              scope: '<версия_API>',
            });

            giga.chat({
                messages: [
                  {
                     role: 'user',
                     content: 'Привет! Расскажи о себе в двух словах'
                  }
                ],
              })
              .then((resp) => {
                console.log(resp.choices[0]?.message.content);
              });
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletion'
            text/event-stream:
              schema:
                type: string
                description: |
                  Событие формата [Server-Sent Events](https://html.spec.whatwg.org/multipage/server-sent-events.html).
                  Каждое событие содерджит сообщение вида `data: <JSON-объект с фрагментом ответа модели>`.
                  В последнем событии приходит сообщение `data: [DONE]`.

                  Пример фрагмента:

                  ```json
                  {
                      "choices": [
                          {
                              "delta": {
                                  "content": "GigaChat is a service capable of interacting with the user in a dialogue format, writing code, and creating texts and images upon the user's request.",
                                  "role": "assistant"
                              },
                              "index": 0
                          }
                      ],
                      "created": 1754637655,
                      "model": "GigaChat:2.0.28.2",
                      "object": "chat.completion"
                  }
                  ```
                # Схема фрагмента ответа модели:
                # $ref: '#/components/schemas/ChatCompletionStream'
              example: |
                data: {"choices":[{"delta":{"content":"GigaChat is a service capable of interacting with the user in a dialogue format, writing code, and creating texts and images upon the user's request.","role":"assistant"},"index":0}],"created":1754637655,"model":"GigaChat:2.0.28.2","object":"chat.completion"}

                data: {"choices":[{"delta":{"content":""},"index":0,"finish_reason":"stop"}],"created":1754637655,"model":"GigaChat:2.0.28.2","object":"chat.completion","usage":{"prompt_tokens":56,"completion_tokens":31,"total_tokens":87,"precached_prompt_tokens":3}}

                data: [DONE]
          description: |
            Успешное выполнение запроса.
            При запуске [потоковой генерации](/ru/gigachat/guides/response-token-streaming) передается с заголовком `Content-Type:	text/event-stream`.
        '400':
          $ref: '#/components/responses/BadRequestFormat'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NoSuchModel'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - Токен доступа: []
      operationId: postChat
      summary: Получить ответ модели на сообщения
      description: |
        Возвращает ответ модели, сгенерированный на основе переданных сообщений.

        Передавайте текст сообщений (поле `content`) в кодировке UTF8.
        Это позволит снизить расход токенов при обработке сообщения.

        При генерации ответа модель может учитывать текстовые документы, изображения и аудиофайлы, сохраненные в хранилище.
        Для этого передайте список идентификаторов файлов в массиве `attachments`.
        При использовании больших текстовых файлов в запросах на генерацию, их содержимое может превышать [размер контекста модели](/ru/gigachat/models#modeli-dlya-generatsii).
        В таком случае вернется [ошибка с кодом 422](/ru/gigachat/api/errors-description?responseCode=422).

        В одном сообщении (объект в массиве `messages`) можно передать только одно изображение.
        В одном запросе можно передать до 10 изображений, независимо от количества сообщений.
        
        :::note
        
        При этом общий размер запроса при работе с изображениями и аудио должен быть меньше 80 Мб.

        Например, ваш запрос может включать текст промпта и идентификаторы изображения размером 12 Мб, и двух аудиофайлов размером 33 Мб и 21 Мб. Что в сумме даст запрос размером больше 66 Мб, в зависимости от размера промпта.

        Размер текстовых документов не влияет на размер запроса, но их содержимое может превышать контекстное окно модели.

        :::

        Подробнее — в разделе [Обработка файлов](/ru/gigachat/guides/working-with-files).

        Запрос на генерацию можно передавать [моделям в раннем доступе](/ru/gigachat/models#obrashenie-k-modelyam-rannego-dostupa).
        К названию модели, которое передается в поле `model`, добавьте постфикс `-preview`.
...
  /embeddings:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingsBody'
      x-codeSamples:
        - lang: "Python"
          label: "gigachat"
          source: |
            """
            Установите библиотеку gigachat с помощью менеджера пакетов pip:

            pip install gigachat
            """
            from gigachat import GigaChat

            giga = GigaChat(
               credentials="<ключ_авторизации>"
            )

            response = giga.embeddings(["Hello world!"])

            print(response)
        - lang: "JavaScript"
          label: "gigachat"
          source: |
            /**
             * Установите библиотеку gigachat с помощью менеджера пакетов npm:
             * 
             * npm install gigachat
             */
            import GigaChat from 'gigachat';

            const giga = new GigaChat({
                credentials: 'ключ_авторизации',
            });

            const response = await giga.embeddings(['Слова слова слова']);

            console.log(response.data);
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Embedding'
          description: OK
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - Токен доступа: []
      operationId: postEmbeddings
      summary: Создать эмбеддинг
      description: |
        Возвращает векторные представления соответствующих текстовых запросов. Индекс объекта с векторным представлением (поле `index`) соответствует индексу строки в массиве `input` запроса.

        Векторное представление выглядит как массив чисел `embedding`. Каждое значение в массиве представляет одну из характеристик или признаков текста, учтенных при вычислении эмбеддинга. Значения образуют числовое представление текста и позволяют анализировать и использовать текст в различных задачах. Как правило, чем ближе значения эмбеддингов друг к другу, тем более семантически близки тексты.

        Для создания эмбеддингов можно использовать [модели Embeddings и EmbeddingsGigaR](/ru/gigachat/models#model-dlya-vektornogo-predstavleniya-teksta).
        Запросы [тарифицируются](/ru/gigachat/api/tariffs) одинаково, независимо от использованной модели.

        :::tip

        Для улучшения результатов при работе с моделью EmbeddingsGigaR следуйте рекомендациям в разделе [Векторное представление текста](/ru/gigachat/guides/embeddings#embeddingsgigar-recommendations).

        :::
...
components:
  parameters:
    xClientId:
      name: X-Client-ID
      in: header
      description: |
        Произвольный идентификатор пользователя, который используется для логирования.

        Если вы передали этот заголовок при запросе на создание изображения, то для скачивания изображения в запросе # START_RAW<APIMethod type="GET" path="/files/{file_id}/content" link="/ru/gigachat/api/reference/rest/get-file-id"/># END_RAW нужно передать этот же заголовок.
      schema:
        type: string
    xRequestId:
      in: header
      name: X-Request-ID
      description: Произвольный идентификатор запроса, который используется для логирования.
      schema:
        type: string
    xSessionId:
      in: header
      name: X-Session-ID
      description: Произвольный идентификатор сессии, который используется для логирования.
      schema:
        type: string
  schemas:
...
    Models:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Model'
        object:
          description: Тип сущности в ответе, например, список.
          type: string
          example: list
...
    File:
      description: Описание файла, доступного в хранилище
      type: object
      properties:
        bytes:
          description: Размер файла в байтах.
          type: integer
          example: 120000
        created_at:
          description: Время создания файла в формате unix timestamp.
          type: integer
          format: unix timestamp
          example: 1677610602
        filename:
          description: Название файла.
          type: string
          example: file123
        id:
          description: |
            Идентификатор файла, который можно использовать при [запросах на генерацию](/ru/gigachat/api/reference/rest/post-chat).
            Для этого идентификаторы нужно передать в массиве `attachments`.
            
            Подробнее — в разделе [Обработка файлов](/ru/gigachat/guides/working-with-files).
          type: string
          format: uuidv4
          pattern: (([0-9a-fA-F-])36)
          example: 6f0b1291-c7f3-43c6-bb2e-9f3efb2dc98e
        object:
          description: Тип объекта.
          type: string
          example: file
        purpose:
          description: Назначение файлов. Значение `general` указывает на то, что файлы могут использоваться для [генерации ответа модели](/ru/gigachat/guides/working-with-files)
          enum:
            - general
          type: string
        access_policy:
          type: string
          description: |
            Доступность файла. Возможные значения:
            
            * `public`;
            * `private`.
          example: private
          default: private
          enum:
            - public
            - private
...
    Files:
      description: Массив объектов с данными доступных файлов.
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/File'
    FileUpload:
      required:
        - file
      type: object
      properties:
        file:
          format: binary
          description: Загружаемый объект.
          type: string
        purpose:
          description: Назначение загружаемого файла.
          enum:
            - general
          type: string
...
    ChatCompletionStream:
      type: object
      description: |
        Сообщение формата [Server-Sent Events](https://html.spec.whatwg.org/multipage/server-sent-events.html).
        Содержит поле `data:` с фрагментом ответа модели.
        В последнем событии приходит сообщение `data: [DONE]`.
      required:
        - data
      properties:
        data:
          type: object
          properties:
            choices:
              type: array
              items:
                type: object
                properties:
                  delta:
                    type: object
                    properties:
                      content:
                        $ref: '#/components/schemas/MessagesRes/properties/content'
                      role:
                        $ref: '#/components/schemas/MessagesRes/properties/role'
                  index:
                    $ref: '#/components/schemas/Choices/properties/index'
                  finish_reason:
                    $ref: '#/components/schemas/Choices/properties/finish_reason'
            created:
              $ref: '#/components/schemas/ChatCompletion/properties/created'
            model:
              $ref: '#/components/schemas/Model/properties/id'
            object:
              $ref: '#/components/schemas/ChatCompletion/properties/object'
            usage:
              $ref: '#/components/schemas/Usage'
...
    ChatCompletion:
      type: object
      properties:
        choices:
          type: array
          description: Массив ответов модели.
          items:
            $ref: '#/components/schemas/Choices'
        created:
          format: unix timestamp
          type: integer
          description: Дата и время создания ответа в формате unix timestamp.
          example: 1678878333
        model:
          $ref: '#/components/schemas/Model/properties/id'
        usage:
          $ref: '#/components/schemas/Usage'
        object:
          type: string
          description: Название вызываемого метода.
          example: chat.completion
...
  responses:
...
    UnauthorizedError:
      description: Ошибка авторизации.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                 type: integer
                 description: HTTP-код сообщения.
                 default: 401
              message:
                type: string
                description: Описание ошибки.
                default: Unauthorized
...
    NoSuchModel:
      description: |
        Указан неверный идентификатор модели.

        Список доступных моделей и их идентификаторов — в разделе [Модели GigaChat](/docs/ru/gigachat/models).
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                 type: integer
                 description: HTTP-код сообщения.
                 default: 404
              message:
                type: string
                description: Описание ошибки.
                default: No such model
    InternalError:
      description: Внутренняя ошибка сервера.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                 type: integer
                 description: HTTP-код сообщения.
                 default: 500
              message:
                type: string
                description: Описание ошибки.
                default: Internal Server Error
    BadRequestFormat:
      description: |
        400 Bad request.

        Некорректный формат запроса.
    TooManyRequests:
      description: Слишком много запросов в единицу времени.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                 type: integer
                 description: HTTP-код сообщения.
                 default: 429
              message:
                type: string
                description: Описание ошибки.
                default: Too many requests
    ValidationError:
      description: Ошибка валидации параметров запроса. Проверьте названия полей и значения параметров.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                 type: integer
                 description: HTTP-код сообщения.
                 default: 422
              message:
                type: string
                description: Описание ошибки.
                example: "Invalid params: repetition_penalty must be in range (0, +inf)"
  securitySchemes:
    Базовая аутентификация:
      description: |
        Базовая (Basic) аутентификация с помощью ключа авторизации — строки, полученной в результате кодирования в base64 идентификатора (Client ID) и клиентского ключа (Client Secret) API.

        Ключ авторизации передается в заголовке `Authorization`, в запросе на [получение токена доступа](/ru/gigachat/api/reference/rest/post-token).

        Как получить ключ авторизации и токен доступа Access token читайте в разделах [Быстрый старт для физических лиц](/ru/gigachat/individuals-quickstart) и [Быстрый старт для ИП и юридических лиц](/ru/gigachat/legal-quickstart).
      scheme: basic
      type: http
    Токен доступа:
      description: Аутентификация с помощью токена доступа Access token. Используется во всех запросах к GigaChat API, кроме запроса на [получение токена доступа](/ru/gigachat/api/reference/rest/post-token).
      type: http
      scheme: bearer
      bearerFormat: JWT